Add files via upload
If you have the same method repeated two or more times in different classes, that's bad. What if you have a 10 line code method and need to repeat it throughout different classes? DRY. don't Repeat Yourself! 

There are two solutions to this: Inheritance and Composition.

Inheritance is a mechanism that allows us to define the common behavior or common functions in one class, and then inherit them in other classes.

We say the Mammal class IS an Animal, so it inherits all the features or all the methods of the Animal Class. In this situation, we refer to each as:
Animal: parent, base class.
Mammal: child, sub class.

This inheritance is not limited to methods. We can also inherit attributes from a parent class.